test: simplify mock implementations for ConfigurableList across tests
