+++ app/shared/components/about-us/foundation-founders/FoundationFounders.test.tsx
@@ -6,16 +6,8 @@ import { createDocNode } from '~/__mocks__/utils';
 import { BLOCK_IDS, PAGE_IDS } from '~/constants/pageBlocks';
 import { TeamMemberWithId } from '~/types/store/pages/about-us/blocks/foundationFounderBlock';
 
-interface MockConfigurableListProps<T> {
-  readonly items: readonly T[];
-  readonly addBtnLabel: string;
-  readonly onCreate: () => void;
-  readonly renderItem: (props: { readonly item: T }) => React.ReactNode;
-  readonly editable: boolean;
-  readonly onDelete: (id: string) => void;
-}
 
-interface MockContributorCardProps {
+export interface MockContributorCardProps {
   readonly contributor: TeamMemberWithId;
   readonly currentLocale: 'uk' | 'en';
 }
@@ -32,36 +24,7 @@ jest.mock('~/shared/hooks/use-page-block/usePageBlock', () => ({
 jest.mock('~/ds-components/collapsible-block/CollapsibleBlock');
 jest.mock('~/ds-components/text-field/TextField');
 
-jest.mock('~/components/configurable-list/ConfigurableList', () => ({
-  __esModule: true,
-  default: <T extends { readonly id: string }>({
-    items,
-    addBtnLabel,
-    onCreate,
-    renderItem,
-    editable,
-    onDelete
-  }: MockConfigurableListProps<T>) => (
-    <div data-testid="configurable-list">
-      {items.map((item) => (
-        <div key={item.id} data-testid={`list-item-${item.id}`}>
-          {renderItem({ item })}
-          {editable && (
-            <button data-testid={`delete-${item.id}`} onClick={() => onDelete(item.id)}>
-              Delete
-            </button>
-          )}
-        </div>
-      ))}
-      {editable && (
-        <button data-testid="add-btn" onClick={onCreate}>
-          {addBtnLabel}
-        </button>
-      )}
-    </div>
-  )
-}));
-
+jest.mock('~/components/configurable-list/ConfigurableList');
 jest.mock('~/components/contributor-card/ContributorCard', () => ({
   __esModule: true,
   ContributorCard: ({ contributor, currentLocale }: MockContributorCardProps) => (
