+++ app/shared/components/privacy-policy/components/points-list/PointsList.test.tsx
@@ -5,45 +5,7 @@ import { PointsList, PointsListProps } from './PointsList';
 import { ListPoint } from '~/shared/hooks/use-points-list/usePointsList';
 
 
-interface MockConfigurableListProps<T> {
-  readonly items: readonly T[];
-  readonly addBtnLabel: string;
-  readonly onCreate: () => void;
-  readonly renderItem: (props: { readonly item: T }) => React.ReactNode;
-  readonly editable: boolean;
-  readonly onDelete: (id: string) => void;
-}
-
-
-jest.mock('~/components/configurable-list/ConfigurableList', () => ({
-  __esModule: true,
-  default: <T extends { readonly id: string }>({
-    items,
-    addBtnLabel,
-    onCreate,
-    renderItem,
-    editable,
-    onDelete
-  }: MockConfigurableListProps<T>) => (
-    <div data-testid="configurable-list">
-      {items.map((item) => (
-        <div key={item.id} data-testid={`list-item-${item.id}`}>
-          {renderItem({ item })}
-          {editable && (
-            <button data-testid={`delete-${item.id}`} onClick={() => onDelete(item.id)}>
-              Delete
-            </button>
-          )}
-        </div>
-      ))}
-      {editable && (
-        <button data-testid="add-btn" onClick={onCreate}>
-          {addBtnLabel}
-        </button>
-      )}
-    </div>
-  )
-}));
+jest.mock('~/components/configurable-list/ConfigurableList');
 
 const defaultProps: PointsListProps = {
   points: [],
